Refactor: Se crea la interfaz Pregunta con sus clases concretas
Se crea la interfaz Pregunta con PreguntaVoFClasica y PreguntaVoFConPenalidad
